requirements.txt
flask==3.0.2
requests==2.31.0
python-dotenv==1.0.1
cryptography==42.0.5
xmltodict==0.13.0
gunicorn==21.2.0
pycryptodome==3.18.0 